+++ scripts/macros/spells/spiritGuardians.js
@@ -61,7 +61,7 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
     if (!effect) return;
     await tokenMove.add('spiritGuardians', castLevel, spellDC, damage, damageType, sourceTokenID, range, true, true, 'start', effect.uuid);
     let color = chris.getConfiguration(workflow.item, 'color') ?? 'blueyellow';
-    let variation = `.${chris.getConfiguration(workflow.item, 'variation') ?? 'ring'}`;
+    let variation = '.' + chris.getConfiguration(workflow.item, 'variation') ?? '.ring';
     if (color === 'random') {
         let colors = [
             'blueyellow',